Added locator regression test
